+++ wicket-examples/src/main/java/org/apache/wicket/examples/repeater/ContactDataProvider.java
@@ -42,7 +42,7 @@ public class ContactDataProvider implements IDataProvider<Contact>
 	 * 
 	 * @see org.apache.wicket.markup.repeater.data.IDataProvider#iterator(int, int)
 	 */
-	public Iterator<Contact> iterator(int first, int count)
+	public Iterator<Contact> iterator(long first, long count)
 	{
 		return getContactsDB().find(first, count, new SortParam("firstName", true)).iterator();
 	}
@@ -52,7 +52,7 @@ public class ContactDataProvider implements IDataProvider<Contact>
 	 * 
 	 * @see org.apache.wicket.markup.repeater.data.IDataProvider#size()
 	 */
-	public int size()
+	public long size()
 	{
 		return getContactsDB().getCount();
 	}
